organized components under layout and pages
