+++ client/src/components/layout/SearchCard/SearchCard.css
