+++ client/src/components/layout/SearchCard/index.js
